+++ README.md
@@ -5,6 +5,7 @@
 [![license](https://img.shields.io/pypi/l/pii-transform)](LICENSE)
 [![build status](https://github.com/piisa/pii-transform/actions/workflows/pii-transform-pr.yml/badge.svg)](https://github.com/piisa/pii-transform/actions)
 
+
 This package takes a source document, a collection of detected PII instances,
 and transforms the document by replacing the PII instances in the document
 with a different representation.
@@ -15,11 +16,12 @@ Note: `pii-transform` does **not** implement or use Transformer models for PII
 purposes (for the extraction of PII Instances using Transformer models, see
 [pii-extract-plg-transformers] or [pii-extract-plg-presidio]).
 
-## Command-line scripts
+## Command-line script
 
-The package provides three console scripts:
+The package provides a console script: `pii-transform` loads a source document
+& a collection of already-detected
 
- * `pii-transform` loads a source document & a collection of already-detected
+
    PII, and produces a transformed document following the required policies.
  * `pii-process` is a full end-to-end script:
     - loads a document, from among the formats supported by `pii-preprocess`
@@ -31,43 +33,22 @@ The package provides three console scripts:
    `[JSONL] files and processes each line as a separate text buffer (possibly in
    different languages), producing a transformed JSONL document
 	  
-	  
-## end-to-end installation
-
-Note that `pii-process` & `pii-process-jsonl` will need additional packages
-to be installed:
- * `pii-preprocess` (only when using `pii-process`)
- * `pii-extract-base`, together with any desired detection plugins, e.g.
-   `pii-extract-plg-regex`, `pii-extract-plg-transformers`,
-   and/or `pii-extract-plg-presidio`
- * `pii-decide`
-
-This installation can be performed explicitly, choosing the packages & plugins
-to install. There is also an automatic dependency installation, which
-installs a default set of packages, by adding the `[e2e]` qualifier upon
-installation of this package, i.e.:
-
-          pip install pii-transform[e2e]
-
-... and this will install `pii_preprocess`, `pii-extract-base`,
-`pii-extract-plg-regex`, `pii-extract-plg-transformers` and `pii-decide`
-
-Note that you will also need to install Pytorch, so that the models used by
-the `pii-extract-plg-transformers` package can run. See the [transformers
-plugin documentation] for more information,
-
 
 ## API
 
-The same functionality provided by the command-line scripts can also be
+The same functionality provided by the command-line script can also be
 accessed via a [Python API]
+
+
+## End-to-end workflow
+
+The end-to-end scripts and APIs have been migrated; they are now in the
+[`pii-process`] package
 
 
 
 [transformation policies]: doc/policies.md
 [Python API]: doc/api.md
-[`pii-process-jsonl`]: doc/jsonl.md
+[`pii-process`]: https://github.com/piisa/pii-process
 [pii-extract-plg-transformers]: https://github.com/piisa/pii-extract-plg-transformers
 [pii-extract-plg-presidio]: https://github.com/piisa/pii-extract-plg-presidio
-[transformers plugin documentation]: https://github.com/piisa/pii-extract-plg-transformers
-[JSONL]: https://jsonlines.org
