+++ src/pii_transform/__init__.py
@@ -1,1 +1,1 @@
-VERSION = "0.6.0"
+VERSION = "0.7.0"
