+++ src/pii_transform/api/e2e.py
@@ -1,0 +1,3 @@
+# Import pii-process API here for backwards compatibility
+from pii_process.api import process_document, PiiTextProcessor, MultiPiiTextProcessor  # noqa: F401
+from .transform import format_policy  # noqa: F401
