--- src/pii_transform/api/e2e/__init__.py
@@ -1,3 +1,0 @@
-from .document import process_document, format_policy, MISSING_MOD  # noqa: F401
-from .textchunk import PiiTextProcessor                             # noqa: F401
-from .multilang import MultiPiiTextProcessor                        # noqa: F401
