--- src/pii_transform/api/e2e/defs.py
@@ -1,3 +1,0 @@
-
-# Minimum vesion  required of pii-extract-base package
-MIN_PII_EXTRACT_VERSION = "0.6.1"
